python/requirements.txt
iphone-backup-decrypt>=0.5.0
Pillow>=10.0
